+++ kernel/arch/x86/drivers/usb_xhci_init.c
@@ -486,7 +486,7 @@ bool _xhci_ensure_running(xhci_controller_t *ctrl) {
                 ctrl->bus,
                 ctrl->slot,
                 ctrl->func,
-                now
+                (unsigned int)now
             );
 
             _xhci_log_fault_snapshot(ctrl, "controller not ready");
@@ -529,7 +529,7 @@ bool _xhci_ensure_running(xhci_controller_t *ctrl) {
             ctrl->bus,
             ctrl->slot,
             ctrl->func,
-            now
+            (unsigned int)now
         );
 
         _xhci_log_fault_snapshot(ctrl, "controller halted");
@@ -553,7 +553,7 @@ bool _xhci_ensure_running(xhci_controller_t *ctrl) {
                 ctrl->bus,
                 ctrl->slot,
                 ctrl->func,
-                now
+                (unsigned int)now
             );
 
             _xhci_log_fault_snapshot(ctrl, "controller never became ready");
@@ -852,7 +852,7 @@ static bool _xhci_validate_runtime_programming(
             ctrl->func,
             dcbaap,
             erstba,
-            erstsz
+            (unsigned int)erstsz
         );
         return false;
     }
@@ -1110,7 +1110,7 @@ static bool _xhci_reset_and_configure(xhci_controller_t *ctrl, bool scan_ports) 
         ctrl->max_slots,
         ctrl->max_ports,
         ctrl->scratchpad_count,
-        hcsp2
+        (unsigned int)hcsp2
     );
 
     u32 hccparams1 = _read32(mmio, XHCI_HCCPARAMS1_OFF);
@@ -1143,7 +1143,7 @@ static bool _xhci_reset_and_configure(xhci_controller_t *ctrl, bool scan_ports) 
             ctrl->bus,
             ctrl->slot,
             ctrl->func,
-            pagesize
+            (unsigned int)pagesize
         );
     }
 
@@ -1199,7 +1199,7 @@ static bool _xhci_reset_and_configure(xhci_controller_t *ctrl, bool scan_ports) 
             ctrl->bus,
             ctrl->slot,
             ctrl->func,
-            now
+            (unsigned int)now
         );
         arch_phys_unmap(map, XHCI_MMIO_SIZE);
         return false;
@@ -1249,7 +1249,7 @@ static bool _xhci_reset_and_configure(xhci_controller_t *ctrl, bool scan_ports) 
             ctrl->bus,
             ctrl->slot,
             ctrl->func,
-            now
+            (unsigned int)now
         );
         arch_phys_unmap(map, XHCI_MMIO_SIZE);
         return false;
@@ -1269,7 +1269,7 @@ static bool _xhci_reset_and_configure(xhci_controller_t *ctrl, bool scan_ports) 
             ctrl->bus,
             ctrl->slot,
             ctrl->func,
-            now
+            (unsigned int)now
         );
         arch_phys_unmap(map, XHCI_MMIO_SIZE);
         return false;
